Finish

+++ laravel/database/migrations/2023_05_24_000733_create_wallet_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('wallet', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('users_id')->nullable();
-            $table->double('balance', 8, 2)->default(0);
+            $table->decimal('balance', 10, 2)->default(0);
             $table->timestamps();
         });
     }
